requirements.txt
google-generativeai
atproto
python-dotenv
schedule
